fix(TCOMP-859): move all parameter resolvers to ...parameter.resolver package
