+++ main/plugins/org.talend.sdk.component.studio-integration/src/main/java/org/talend/sdk/component/studio/model/parameter/resolver/ValidationResolver.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.talend.sdk.component.studio.model.parameter;
+package org.talend.sdk.component.studio.model.parameter.resolver;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -26,9 +26,11 @@ import org.talend.sdk.component.server.front.model.ActionReference;
 import org.talend.sdk.component.server.front.model.SimplePropertyDefinition;
 import org.talend.sdk.component.studio.model.action.Action;
 import org.talend.sdk.component.studio.model.action.ActionParameter;
+import org.talend.sdk.component.studio.model.parameter.PropertyNode;
+import org.talend.sdk.component.studio.model.parameter.TaCoKitElementParameter;
 import org.talend.sdk.component.studio.model.parameter.listener.ValidationListener;
 
-class ValidationResolver extends AbstractParameterResolver {
+public class ValidationResolver extends AbstractParameterResolver {
 
     private final ActionReference action;
 
@@ -36,7 +38,7 @@ class ValidationResolver extends AbstractParameterResolver {
 
     private final ElementParameter redrawParameter;
 
-    ValidationResolver(final PropertyNode actionOwner, final Collection<ActionReference> actions,
+    public ValidationResolver(final PropertyNode actionOwner, final Collection<ActionReference> actions,
             final ValidationListener listener, final ElementParameter redrawParameter) {
         super(actionOwner);
         if (!actionOwner.getProperty().hasValidation()) {
